# s3_manda2-web/target/jaxrs-analyzer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "manda2-web"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/organizacion": {
            "get": {
                "description": "<h1>GET /api/organizaicin: Obtener todos los servicios de tipo Organización.</h1>\n \n <pre>Busca y devuelve todos los servicios de tipo Organización que existen en la aplicación.\n \n Códigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK Devuelve todos los servicios de la aplicacion.</code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/OrganizacionDetailsDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "<h1>POST /api/organizacion: Crear un servicio de tipo Organización.</h1>\n \n <pre>Cuerpo de petición: JSON {@link servicioDTO}.\n \n Crea un nuevo servicio de tipo Organización con la información que se recibe en el cuerpo de la petición\n y se regresa un objeto idéntico con un id auto-generado por la base de datos.\n \n Códigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK Creó el nuevo servicio de tipo Organización.\n </code>\n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 412 Precodition Failed: Ya existe el servicio de tipo Organización.\n </code>\n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OrganizacionDetailsDTO"
                        },
                        "description": "{@link ServicioDetailDTO} - El servicio de tipo Organización que se desea guardar."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/OrganizacionDetailsDTO"
                        }
                    }
                }
            }
        },
        "/organizacion/{id }": {
            "get": {
                "description": "<h1>GET /api/organizacion/{id} : Obtener servicio de organizacion por id.</h1>\n \n <pre>Busca el servicio con el id asociado recibido en la URL y la devuelve.\n \n Codigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK Devuelve el servicio de tipo organización correspondiente al id.\n </code> \n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 404 Not Found No existe un servicio de tipo organización con el id dado.\n </code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identificador del servicio de tipo organización que se está buscando. Este debe ser una cadena de dígitos."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/OrganizacionDetailsDTO"
                        }
                    }
                }
            },
            "put": {
                "description": "/**\n <h1>PUT /api/organizacion/{id} : Actualizar servicio de tipo Organización con el id dado.</h1>\n <pre>Cuerpo de petición: JSON {@link CityDetailDTO}.\n \n Actualiza el servicio de tipo Organización con el id recibido en la URL con la informacion que se recibe en el cuerpo de la petición.\n \n Codigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK Actualiza el servicio de tipo Organización con el id dado con la información enviada como parámetro. Retorna un objeto identico.</code> \n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 404 Not Found. No existe un servicio de tipo Organización con el id dado.\n </code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identificador del servicio de tipo Organización que se desea actualizar. Este debe ser una cadena de dígitos."
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OrganizacionDetailsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/OrganizacionDetailsDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "<h1>DELETE /api/organizacion/{id} : Borrar servicio de tipo Organización por id.</h1>\n \n <pre>Borra el servicio de tipo Organización con el id asociado recibido en la URL.\n \n Códigos de respuesta:<br>\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK Elimina el servicio de tipo Organización correspondiente al id dado.</code>\n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 404 Not Found. No existe un servicio de tipo Organización con el id dado.\n </code>\n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identificador del servicio de tipo Organización que se desea borrar. Este debe ser una cadena de dígitos."
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OrganizacionDetailsDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "OrganizacionDTO": {
            "properties": {
                "costoDuracion": {
                    "type": "number"
                },
                "costoTransporte": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "OrganizacionDetailsDTO": {
            "properties": {
                "costoDuracion": {
                    "type": "number"
                },
                "costoTransporte": {
                    "type": "number"
                },
                "elementoBusqueda": {
                    "$ref": "#/definitions/OrganizacionDTO"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
    }
}